requirements.txt
fastapi==0.115.2
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.7.4
alembic==1.12.1
redis==5.0.1
pytest==7.4.3
python-dotenv==1.0.0
httpx==0.25.2 